Reorganize foldrs
