+++ example/src/camera/CameraTransitionManager.js
