+++ example/src/plugins/fade/FadeManager.js
