+++ example/src/plugins/fade/FadeTilesRenderer.js
