+++ example/src/plugins/overlays/TextureOverlayMaterial.js
